slackbots/botlib-bots/bot-lib/requirements.txt
ruamel.yaml==0.11.7
pyOpenSSL==17.5.0
requests>=2.20.0
pandas==0.21.1
MySQL-python==1.2.5
slackclient==1.3.2
simple-rbac==0.1.1
sshtunnel==0.1.2
pyyaml
#jira
APScheduler==3.5.0
xhtml2pdf==0.2b1
pytablewriter==0.27.2
sops==1.16
singleton-decorator==1.0.0
dateparser==0.6.0
